+++ widgets/status.js
@@ -5,12 +5,12 @@ $.widget("sv.status_badge", $.sv.widget, {
 
 	options: {
 		 thresholds: '',
-		 colors: ''
+		 colors: '',
+		 symbols: ''
 	},
 	
 	_update: function(response) {
-		this.element.children('span').text(response[0]);
-
+		
 		// coloring
 		var currentIndex = 0;
 		$.each(String(this.options.thresholds).explode(), function(index, threshold) {
@@ -19,6 +19,14 @@ $.widget("sv.status_badge", $.sv.widget, {
 			currentIndex++;
 		});
 		var color = String(this.options.colors).explode()[currentIndex];
+		var symbol = String(this.options.symbols).explode()[currentIndex] || '';
+
+		if (symbol == '')
+			this.element.children('span').text(response[0]);
+		else {
+			this.element.children('span').text('');
+			fx.load(symbol, 'icon', 'stroke: #fff; fill: #fff', this.element.children('span'), 'append'  )
+		}
 
 		if(color == 'hidden') {
 			this.element.children('span').hide().css('background-color', null);