+++ changelog.md
@@ -6,6 +6,7 @@
 - device.uzsuicon and device.uzsugraph support the "perday" option for interpolation (interpolation with the days events only, not across day borders)
 - Enertex (R) appliance.iprouter_v2 widget has been adapted to older device types and has got a new option to hide closed tunnels (thanks to @stoepf)
 - status.collapse and status.customstyle allow items of type "str" (no comparative operators (<,<=,=>,>) allowed with text). 
+- status badge is now able to display symbols
 
 ### Other New Features